Initial deposit
Initial version of Globus transfer management service.
Currently expects script to be executed from a personal directory & requires a Globus personal instance to be running. Will update in -dev version once we have globus server running.
Projects to check for on upload server controlled by projects.txt which on deployment is a link to which ever projects_XXXX.txtx file is required.